// src/types/ChromeMessageTypes.ts
export enum ChromeMessageTypes {
  QUEUE_STATUS = "QUEUE_STATUS",
  SEND_MESSAGE = "SEND_MESSAGE",
  PAUSE_QUEUE = "PAUSE_QUEUE",
  RESUME_QUEUE = "RESUME_QUEUE",
  STOP_QUEUE = "STOP_QUEUE",
  ADD_LOG = "ADD_LOG",
  ARCHIVE_ALL_CHATS = "ARCHIVE_ALL_CHATS",
  ARCHIVE_STATUS = "ARCHIVE_STATUS",
  WAJS_LAB_EXECUTE = "WAJS_LAB_EXECUTE",
  SCHEDULE_EXECUTION = "SCHEDULE_EXECUTION",
  LIST_SCHEDULED_EXECUTIONS = "LIST_SCHEDULED_EXECUTIONS",
  CANCEL_SCHEDULED_EXECUTION = "CANCEL_SCHEDULED_EXECUTION",
}